refactoring profile screen
